EZP-24664: removed main search handler interface
