+++ eZ/Publish/Core/Repository/Repository.php
@@ -800,7 +800,7 @@ class Repository implements RepositoryInterface
 
         $this->searchService = new SearchService(
             $this,
-            $this->searchHandler->contentSearchHandler(),
+            $this->searchHandler,
             $this->getDomainMapper(),
             $this->getPermissionsCriterionHandler(),
             $this->serviceSettings['search']
