+++ eZ/Publish/Core/Repository/SearchService.php
@@ -23,7 +23,7 @@ use eZ\Publish\API\Repository\Values\Content\Search\SearchResult;
 use eZ\Publish\Core\Base\Exceptions\NotFoundException;
 use eZ\Publish\Core\Base\Exceptions\InvalidArgumentException;
 use eZ\Publish\Core\Base\Exceptions\InvalidArgumentType;
-use eZ\Publish\SPI\Search\Content\Handler;
+use eZ\Publish\SPI\Search\Handler;
 
 /**
  * Search service.
@@ -36,7 +36,7 @@ class SearchService implements SearchServiceInterface
     protected $repository;
 
     /**
-     * @var \eZ\Publish\SPI\Search\Content\Handler
+     * @var \eZ\Publish\SPI\Search\Handler
      */
     protected $searchHandler;
 
@@ -59,7 +59,7 @@ class SearchService implements SearchServiceInterface
      * Setups service with reference to repository object that created it & corresponding handler.
      *
      * @param \eZ\Publish\API\Repository\Repository $repository
-     * @param \eZ\Publish\SPI\Search\Content\Handler $searchHandler
+     * @param \eZ\Publish\SPI\Search\Handler $searchHandler
      * @param \eZ\Publish\Core\Repository\Helper\DomainMapper $domainMapper
      * @param \eZ\Publish\Core\Repository\PermissionsCriterionHandler $permissionsCriterionHandler
      * @param array $settings
