+++ eZ/Publish/Core/Search/Common/Slot/DeleteLocation.php
@@ -29,7 +29,7 @@ class DeleteLocation extends Slot
             return;
         }
 
-        $this->searchHandler->contentSearchHandler()->deleteLocation(
+        $this->searchHandler->deleteLocation(
             $signal->locationId,
             $signal->contentId
         );
