+++ eZ/Publish/Core/Search/Common/Slot/DeleteVersion.php
@@ -29,7 +29,7 @@ class DeleteVersion extends Slot
             return;
         }
 
-        $this->searchHandler->contentSearchHandler()->deleteContent(
+        $this->searchHandler->deleteContent(
             $signal->contentId,
             $signal->versionNo
         );
