+++ eZ/Publish/Core/Search/Common/Slot/HideLocation.php
@@ -29,14 +29,14 @@ class HideLocation extends Slot
             return;
         }
 
-        $this->searchHandler->contentSearchHandler()->indexContent(
+        $this->searchHandler->indexContent(
             $this->persistenceHandler->contentHandler()->load(
                 $signal->contentId,
                 $signal->currentVersionNo
             )
         );
 
-        $this->searchHandler->contentSearchHandler()->indexLocation(
+        $this->searchHandler->indexLocation(
             $this->persistenceHandler->locationHandler()->load($signal->locationId)
         );
     }
