+++ eZ/Publish/Core/Search/Common/Slot/MoveSubtree.php
@@ -35,14 +35,13 @@ class MoveSubtree extends Slot
     protected function indexSubtree($locationId)
     {
         $contentHandler = $this->persistenceHandler->contentHandler();
-        $contentSearchHandler = $this->searchHandler->contentSearchHandler();
         $locationHandler = $this->persistenceHandler->locationHandler();
 
         $processedContentIdSet = array();
         $subtreeIds = $locationHandler->loadSubtreeIds($locationId);
 
         foreach ($subtreeIds as $locationId => $contentId) {
-            $contentSearchHandler->indexLocation(
+            $this->searchHandler->indexLocation(
                 $locationHandler->load($locationId)
             );
 
@@ -50,7 +49,7 @@ class MoveSubtree extends Slot
                 continue;
             }
 
-            $contentSearchHandler->indexContent(
+            $this->searchHandler->indexContent(
                 $contentHandler->load(
                     $contentId,
                     $contentHandler->loadContentInfo($contentId)->currentVersionNo
