+++ eZ/Publish/Core/Search/Common/Slot/SetContentState.php
@@ -31,7 +31,7 @@ class SetContentState extends Slot
 
         $contentInfo = $this->persistenceHandler->contentHandler()->loadContentInfo($signal->contentId);
 
-        $this->searchHandler->contentSearchHandler()->indexContent(
+        $this->searchHandler->indexContent(
             $this->persistenceHandler->contentHandler()->load(
                 $contentInfo->id,
                 $contentInfo->currentVersionNo
@@ -40,7 +40,7 @@ class SetContentState extends Slot
 
         $locations = $this->persistenceHandler->locationHandler()->loadLocationsByContent($contentInfo->id);
         foreach ($locations as $location) {
-            $this->searchHandler->contentSearchHandler()->indexLocation($location);
+            $this->searchHandler->indexLocation($location);
         }
     }
 }
