+++ eZ/Publish/Core/Search/Common/Slot/UnhideLocation.php
@@ -29,14 +29,14 @@ class UnhideLocation extends Slot
             return;
         }
 
-        $this->searchHandler->contentSearchHandler()->indexContent(
+        $this->searchHandler->indexContent(
             $this->persistenceHandler->contentHandler()->load(
                 $signal->contentId,
                 $signal->currentVersionNo
             )
         );
 
-        $this->searchHandler->contentSearchHandler()->indexLocation(
+        $this->searchHandler->indexLocation(
             $this->persistenceHandler->locationHandler()->load($signal->locationId)
         );
     }
