+++ eZ/Publish/Core/settings/search_engines/elasticsearch.yml
@@ -13,12 +13,11 @@ parameters:
     ezpublish.search.elasticsearch.connection.index_name: 'ezpublish'
     ezpublish.search.elasticsearch.connection.document_type_identifier.content: 'content'
     ezpublish.search.elasticsearch.connection.document_type_identifier.location: 'location'
-    ezpublish.spi.search.elasticsearch.class: eZ\Publish\Core\Search\Elasticsearch\Handler
+    ezpublish.spi.search.elasticsearch.class: eZ\Publish\Core\Search\Elasticsearch\Content\Handler
     ezpublish.search.elasticsearch.serializer.class: eZ\Publish\Core\Search\Elasticsearch\Content\Serializer
     ezpublish.search.elasticsearch.mapper.standard.class: eZ\Publish\Core\Search\Elasticsearch\Content\Mapper\StandardMapper
     ezpublish.search.elasticsearch.gateway.native.class: eZ\Publish\Core\Search\Elasticsearch\Content\Gateway\Native
     ezpublish.search.elasticsearch.location.gateway.native.class: eZ\Publish\Core\Search\Elasticsearch\Content\Location\Gateway\Native
-    ezpublish.spi.search.elasticsearch.content_handler.class: eZ\Publish\Core\Search\Elasticsearch\Content\Handler
     ezpublish.search.elasticsearch.extractor.loading.class: eZ\Publish\Core\Search\Elasticsearch\Content\Extractor\Loading
 
 services:
@@ -78,8 +77,8 @@ services:
     ezpublish.search.elasticsearch.extractor:
         alias: ezpublish.search.elasticsearch.extractor.loading
 
-    ezpublish.spi.search.elasticsearch.content_handler:
-        class: %ezpublish.spi.search.elasticsearch.content_handler.class%
+    ezpublish.spi.search.elasticsearch:
+        class: %ezpublish.spi.search.elasticsearch.class%
         arguments:
             - @ezpublish.search.elasticsearch.content.gateway
             - @ezpublish.search.elasticsearch.location.gateway
@@ -87,12 +86,6 @@ services:
             - @ezpublish.search.elasticsearch.extractor
             - %ezpublish.search.elasticsearch.connection.document_type_identifier.content%
             - %ezpublish.search.elasticsearch.connection.document_type_identifier.location%
-        lazy: true
-
-    ezpublish.spi.search.elasticsearch:
-        class: %ezpublish.spi.search.elasticsearch.class%
-        arguments:
-            - @ezpublish.spi.search.elasticsearch.content_handler
         tags:
             - {name: ezpublish.searchEngine, alias: elasticsearch}
         lazy: true
