+++ eZ/Publish/Core/settings/search_engines/legacy.yml
@@ -8,9 +8,7 @@ imports:
     - {resource: search_engines/legacy/sort_clause_handlers_location.yml}
 
 parameters:
-    ezpublish.spi.search.legacy.class: eZ\Publish\Core\Search\Legacy\Handler
-    ezpublish.spi.search.legacy.content_handler.class: eZ\Publish\Core\Search\Legacy\Content\Handler
-    ezpublish.spi.search.legacy.location_handler.class: eZ\Publish\Core\Search\Legacy\Content\Location\Handler
+    ezpublish.spi.search.legacy.class: eZ\Publish\Core\Search\Legacy\Content\Handler
     ezpublish.search.legacy.gateway.content.class: eZ\Publish\Core\Search\Legacy\Content\Gateway\DoctrineDatabase
     ezpublish.search.legacy.gateway.content.exception_conversion.class: eZ\Publish\Core\Search\Legacy\Content\Gateway\ExceptionConversion
     ezpublish.search.legacy.gateway.location.class: eZ\Publish\Core\Search\Legacy\Content\Location\Gateway\DoctrineDatabase
@@ -66,19 +64,13 @@ services:
     ezpublish.search.legacy.gateway.location:
         alias: ezpublish.search.legacy.gateway.location.exception_conversion
 
-    ezpublish.spi.search.legacy.content_handler:
-        class: %ezpublish.spi.search.legacy.content_handler.class%
+    ezpublish.spi.search.legacy:
+        class: %ezpublish.spi.search.legacy.class%
         arguments:
             - @ezpublish.search.legacy.gateway.content
             - @ezpublish.search.legacy.gateway.location
             - @ezpublish.persistence.legacy.content.mapper
             - @ezpublish.persistence.legacy.location.mapper
-        lazy: true
-
-    ezpublish.spi.search.legacy:
-        class: %ezpublish.spi.search.legacy.class%
-        arguments:
-            - @ezpublish.spi.search.legacy.content_handler
         tags:
             - {name: ezpublish.searchEngine, alias: legacy}
         lazy: true
