--- eZ/Publish/SPI/Search/Content/Handler.php
@@ -1,103 +1,0 @@
-<?php
-
-/**
- * File containing the Content Search handler interface.
- *
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- *
- * @version //autogentag//
- */
-namespace eZ\Publish\SPI\Search\Content;
-
-use eZ\Publish\SPI\Persistence\Content;
-use eZ\Publish\SPI\Persistence\Content\Location;
-use eZ\Publish\API\Repository\Values\Content\Query\Criterion;
-use eZ\Publish\API\Repository\Values\Content\Query;
-use eZ\Publish\API\Repository\Values\Content\LocationQuery;
-
-/**
- * The Content Search handler retrieves sets of of Content objects, based on a
- * set of criteria.
- */
-interface Handler
-{
-    /**
-     * Finds content objects for the given query.
-     *
-     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if Query criterion is not applicable to its target
-     *
-     * @param \eZ\Publish\API\Repository\Values\Content\Query $query
-     * @param array $languageFilter Configuration for specifying prioritized languages query will be performed on.
-     *        Currently supported: <code>array("languages" => array(<language1>,..))</code>.
-     *
-     * @return \eZ\Publish\API\Repository\Values\Content\Search\SearchResult With ContentInfo as SearchHit->valueObject
-     */
-    public function findContent(Query $query, array $languageFilter = array());
-
-    /**
-     * Performs a query for a single content object.
-     *
-     * @throws \eZ\Publish\API\Repository\Exceptions\NotFoundException if the object was not found by the query or due to permissions
-     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if Criterion is not applicable to its target
-     * @throws \eZ\Publish\API\Repository\Exceptions\InvalidArgumentException if there is more than than one result matching the criterions
-     *
-     * @param \eZ\Publish\API\Repository\Values\Content\Query\Criterion $filter
-     * @param array $languageFilter Configuration for specifying prioritized languages query will be performed on.
-     *        Currently supported: <code>array("languages" => array(<language1>,..))</code>.
-     *
-     * @return \eZ\Publish\SPI\Persistence\Content\ContentInfo
-     */
-    public function findSingle(Criterion $filter, array $languageFilter = array());
-
-    /**
-     * Finds locations for the given $query.
-     *
-     * @param \eZ\Publish\API\Repository\Values\Content\LocationQuery $query
-     * @param array $fieldFilters - a map of filters for the returned fields.
-     *        Currently supported: <code>array("languages" => array(<language1>,..))</code>.
-     *
-     * @return \eZ\Publish\API\Repository\Values\Content\Search\SearchResult With Location as SearchHit->valueObject
-     */
-    public function findLocations(LocationQuery $query, array $fieldFilters = array());
-
-    /**
-     * Suggests a list of values for the given prefix.
-     *
-     * @param string $prefix
-     * @param string[] $fieldPaths
-     * @param int $limit
-     * @param \eZ\Publish\API\Repository\Values\Content\Query\Criterion $filter
-     */
-    public function suggest($prefix, $fieldPaths = array(), $limit = 10, Criterion $filter = null);
-
-    /**
-     * Indexes a content object.
-     *
-     * @param \eZ\Publish\SPI\Persistence\Content $content
-     */
-    public function indexContent(Content $content);
-
-    /**
-     * Deletes a content object from the index.
-     *
-     * @param int $contentId
-     * @param int|null $versionId
-     */
-    public function deleteContent($contentId, $versionId = null);
-
-    /**
-     * Indexes a Location in the index storage.
-     *
-     * @param \eZ\Publish\SPI\Persistence\Content\Location $location
-     */
-    public function indexLocation(Location $location);
-
-    /**
-     * Deletes a location from the index.
-     *
-     * @param mixed $locationId
-     * @param mixed $contentId
-     */
-    public function deleteLocation($locationId, $contentId);
-}
